+++ src/spotted/utils/__init__.py
@@ -10,3 +10,4 @@ from .keyboard_util import (
     get_published_post_kb,
     get_settings_kb,
 )
+from .utils import get_user_by_id_or_index